Rework unary mutation and add tests
Now, whenever we find an unary operator inside a binary expression
or as a function call argument, we replace them with the expression
its holding inside and we create a mutant.

Previously, we swapped the unary operator found with one of the
remaining two, but that almost always led to a syntax error when
running the mutants, so it was pointless.